The Fill-less Code Runs!!!
The code runs and produces correct results for simple test fill
at the bottom of dft_newton.c. This verifies the Newton algorithm,
the hooks to the dft_SolverManager, and the linear solver. The
mesh, guess, and ouput routines are running but this test problem
does not do much to verify that they are running correctly.

The code runs in parallel as well, giving the exact same result
to all printed digits in the dft_dens.dat file. I found this
suspicious but the test problem is simple enough that I now
believe it is really running in parallel.

To get this to run, I edited the Makefile to not compile
any dft_fill* file except for dft_filled_shared.c, and to
also skip dft_continuation.c and loca*. I also added configure
and link paths to my installed Trilinos code.

I am adding dft_SolverManager.* dft_c2cpp_wrappers.* to this
source directory, just direct copies from the now-old NEW directory.

+++ src/dft_comm.c
@@ -22,19 +22,19 @@ double gmin_double(double c){
   return r;
 }
 /***************************************************************************/
-double gsum_int(int c){
+int gsum_int(int c){
   int r;
   (void) MPI_Allreduce(&c, &r, 1,  MPI_INT, MPI_SUM, Comm);
   return r;
 }
 /***************************************************************************/
-double gmax_int(int c){
+int gmax_int(int c){
   int r;
   (void) MPI_Allreduce(&c, &r, 1,  MPI_INT, MPI_MAX, Comm);
   return r;
 }
 /***************************************************************************/
-double gmin_int(int c){
+int gmin_int(int c){
   int r;
   (void) MPI_Allreduce(&c, &r, 1,  MPI_INT, MPI_MIN, Comm);
   return r;
